requirements.txt
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.3
filelock==3.3.2
idna==3.3
platformdirs==2.4.0
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
soupsieve==2.3
urllib3==1.26.7
virtualenv==20.10.0
vk-api==11.9.6
